pyproject.toml
[tool.poetry]
name = "pdf-to-text"
version = "0.1.0"
description = ""
authors = ["Earl Hikky <earl.hikky91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pdfminer-six = "^20221114"
genanki = "^0.13.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
